// frontend/src/lib/types/generated/RunConclusion.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Conclusion of a completed workflow run.
 */
export type RunConclusion = "Success" | "Failure" | "Cancelled" | "TimedOut" | "ActionRequired" | "Stale" | "Neutral" | "Skipped" | "StartupFailure";